+++ apps/demo-react-cra/src/App.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import logo from "./logo.svg";
 import "./App.css";
-import * as asm from "@fuels/vm-asm";
 import { BaseAssetId, encrypt, decrypt } from "fuels";
+import * as asm from "@fuels/vm-asm";
 
 function App() {
   const { log } = console;