+++ apps/docs-snippets/src/guide/contracts/calls-with-different-wallets.test.ts
@@ -11,7 +11,7 @@ describe(__filename, () => {
     deployedContract = await createAndDeployContractFromProject(SnippetProjectEnum.RETURN_CONTEXT);
   });
 
-  it('should successfully update contract instace wallet', () => {
+  it('should successfully update contract instance wallet', () => {
     const newWallet = WalletUnlocked.generate();
 
     expect(deployedContract.account?.address).not.toBe(newWallet.address);
@@ -23,7 +23,7 @@ describe(__filename, () => {
     expect(deployedContract.account.address).toBe(newWallet.address);
   });
 
-  it('should successfully update contract instace provider', () => {
+  it('should successfully update contract instance provider', () => {
     const newProvider = new Provider('http://provider:9999');
 
     expect(deployedContract.provider?.url).not.toBe(newProvider.url);