+++ packages/abi-coder/src/function-fragment.ts
@@ -2,7 +2,7 @@ import type { BytesLike } from '@ethersproject/bytes';
 import { arrayify } from '@ethersproject/bytes';
 import { Logger } from '@ethersproject/logger';
 import { sha256 } from '@ethersproject/sha2';
-import { bufferFromString } from '@fuel-ts/keystore';
+import { bufferFromString } from '@fuel-ts/crypto';
 import { bn } from '@fuel-ts/math';
 import { versions } from '@fuel-ts/versions';
 
@@ -11,13 +11,14 @@ import type { DecodedValue, InputValue } from './coders/abstract-coder';
 import type { ArrayCoder } from './coders/array';
 import { TupleCoder } from './coders/tuple';
 import type { U64Coder } from './coders/u64';
-import { arrayRegEx, enumRegEx, OPTION_CODER_TYPE, stringRegEx, structRegEx } from './constants';
+import { OPTION_CODER_TYPE } from './constants';
 import type {
   JsonAbi,
   JsonAbiArgument,
   JsonAbiFunction,
   JsonAbiFunctionAttribute,
 } from './json-abi';
+import { ResolvedAbiType } from './resolved-abi-type';
 import type { Uint8ArrayWithDynamicData } from './utilities';
 import { isPointerType, unpackDynamicData, findOrThrow } from './utilities';
 
@@ -37,7 +38,7 @@ export class FunctionFragment<
 
   constructor(abi: JsonAbi, name: string) {
     this.jsonAbi = abi;
-    this.jsonFn = findOrThrow(abi.functions, (f) => f.name === name);
+    this.jsonFn = findOrThrow(this.jsonAbi.functions, (f) => f.name === name);
     this.name = name;
     this.signature = FunctionFragment.getSignature(this.jsonAbi, this.jsonFn);
     this.selector = FunctionFragment.getFunctionSelector(this.signature);
@@ -46,63 +47,10 @@ export class FunctionFragment<
   }
 
   private static getSignature(abi: JsonAbi, fn: JsonAbiFunction): string {
-    const inputsSignatures = fn.inputs.map((input) => this.getArgSignature(abi, input));
+    const inputsSignatures = fn.inputs.map((input) =>
+      new ResolvedAbiType(abi, input).getSignature()
+    );
     return `${fn.name}(${inputsSignatures.join(',')})`;
-  }
-
-  private static getArgSignature(abi: JsonAbi, arg: JsonAbiArgument): string {
-    const prefix = this.getArgSignaturePrefix(abi, arg);
-    const content = this.getArgSignatureContent(abi, arg);
-
-    return `${prefix}${content}`;
-  }
-
-  private static getArgSignaturePrefix(abi: JsonAbi, input: JsonAbiArgument): string {
-    const abiType = findOrThrow(abi.types, (x) => x.typeId === input.type);
-    const structMatch = structRegEx.test(abiType.type);
-    if (structMatch) return 's';
-
-    const arrayMatch = arrayRegEx.test(abiType.type);
-    if (arrayMatch) return 'a';
-
-    const enumMatch = enumRegEx.test(abiType.type);
-    if (enumMatch) return 'e';
-
-    return '';
-  }
-
-  private static getArgSignatureContent(abi: JsonAbi, input: JsonAbiArgument): string {
-    const abiType = findOrThrow(abi.types, (x) => x.typeId === input.type);
-
-    if (abiType.type === 'raw untyped ptr') {
-      return 'rawptr';
-    }
-
-    const strMatch = stringRegEx.exec(abiType.type)?.groups;
-    if (strMatch) {
-      return `str[${strMatch.length}]`;
-    }
-
-    let components = abiType.components;
-
-    if (components === null) return abiType.type;
-
-    components = AbiCoder.resolveGenericComponents(abi, input);
-
-    const arrayMatch = arrayRegEx.exec(abiType.type)?.groups;
-
-    if (arrayMatch) {
-      return `[${this.getArgSignature(abi, components[0])};${arrayMatch.length}]`;
-    }
-
-    const typeArgumentsSignature = Array.isArray(input.typeArguments)
-      ? `<${input.typeArguments.map((arg) => this.getArgSignature(abi, arg)).join(',')}>`
-      : '';
-    const componentsSignature = `(${components
-      .map((arg) => this.getArgSignature(abi, arg))
-      .join(',')})`;
-
-    return `${typeArgumentsSignature}${componentsSignature}`;
   }
 
   private static getFunctionSelector(functionSignature: string) {
@@ -138,16 +86,16 @@ export class FunctionFragment<
 
     const shallowCopyValues = inputValuesArray.slice();
 
-    const nonEmptyTypes = this.jsonFn.inputs.filter(
+    const nonEmptyInputs = this.jsonFn.inputs.filter(
       (x) => findOrThrow(this.jsonAbi.types, (t) => t.typeId === x.type).type !== '()'
     );
 
-    if (Array.isArray(inputValuesArray) && nonEmptyTypes.length !== inputValuesArray.length) {
+    if (Array.isArray(inputValuesArray) && nonEmptyInputs.length !== inputValuesArray.length) {
       shallowCopyValues.length = this.jsonFn.inputs.length;
       shallowCopyValues.fill(undefined as unknown as InputValue, inputValuesArray.length);
     }
 
-    const coders = nonEmptyTypes.map((type) => AbiCoder.getCoder(this.jsonAbi, type));
+    const coders = nonEmptyInputs.map((t) => AbiCoder.getCoder(this.jsonAbi, t));
 
     const coder = new TupleCoder(coders);
     const results: Uint8ArrayWithDynamicData = coder.encode(shallowCopyValues);
@@ -174,11 +122,11 @@ export class FunctionFragment<
 
   decodeArguments(data: BytesLike) {
     const bytes = arrayify(data);
-    const nonEmptyTypes = this.jsonFn.inputs.filter(
+    const nonEmptyInputs = this.jsonFn.inputs.filter(
       (x) => findOrThrow(this.jsonAbi.types, (t) => t.typeId === x.type).type !== '()'
     );
 
-    if (nonEmptyTypes.length === 0) {
+    if (nonEmptyInputs.length === 0) {
       // The VM is current return 0x0000000000000000, but we should treat it as undefined / void
       if (bytes.length === 0) return undefined;
 
@@ -188,19 +136,19 @@ export class FunctionFragment<
         {
           count: {
             types: this.jsonFn.inputs.length,
-            nonEmptyTypes: nonEmptyTypes.length,
+            nonEmptyInputs: nonEmptyInputs.length,
             values: bytes.length,
           },
           value: {
             args: this.jsonFn.inputs,
-            nonEmptyTypes,
+            nonEmptyInputs,
             values: bytes,
           },
         }
       );
     }
 
-    const result = nonEmptyTypes.reduce(
+    const result = nonEmptyInputs.reduce(
       (obj: { decoded: unknown[]; offset: number }, input, currentIndex) => {
         const coder = AbiCoder.getCoder(this.jsonAbi, input);
         if (currentIndex === 0) {