+++ packages/abi-coder/test/interface.test.ts
@@ -411,6 +411,45 @@ describe('Abi interface', () => {
         //   ],
         // },
         {
+          fn: exhaustiveExamplesInterface.functions.array_with_generic_struct,
+          title: '[array] with generic struct',
+          value: [
+            {
+              a: [
+                {
+                  bim: B256_DECODED,
+                  bam: { propB1: U8_MAX },
+                  bom: { propA1: U8_MAX, propA2: B256_DECODED },
+                },
+                {
+                  bim: B256_DECODED,
+                  bam: { propB1: 0 },
+                  bom: { propA1: U8_MAX, propA2: B256_DECODED },
+                },
+                {
+                  bim: B256_DECODED,
+                  bam: { propB1: U8_MAX },
+                  bom: { propA1: U8_MAX, propA2: B256_DECODED },
+                },
+              ],
+            },
+          ],
+          encodedValue: [
+            B256_ENCODED,
+            U8_MAX_ENCODED,
+            U8_MAX_ENCODED,
+            B256_ENCODED,
+            B256_ENCODED,
+            EMPTY_U8_ARRAY,
+            U8_MAX_ENCODED,
+            B256_ENCODED,
+            B256_ENCODED,
+            U8_MAX_ENCODED,
+            U8_MAX_ENCODED,
+            B256_ENCODED,
+          ],
+        },
+        {
           fn: exhaustiveExamplesInterface.functions.vector_boolean,
           title: '[vector] boolean',
           value: { x: [true, false, true, true] },
@@ -654,9 +693,9 @@ describe('Abi interface', () => {
         },
       ])(
         '$title: $value',
-        ({ fn, title: _, value, encodedValue, decodedTransformer, skipDecoding }) => {
+        ({ fn, title: _title, value, encodedValue, decodedTransformer, skipDecoding, offset }) => {
           // @ts-expect-error value is an intersection of all parameterized tests and it's breaking TS
-          const encoded = fn.encodeArguments(value);
+          const encoded = fn.encodeArguments(value, offset);
 
           const encodedVal = encodedValue instanceof Function ? encodedValue(value) : encodedValue;
           const expectedEncoded =