+++ packages/crypto/src/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './scrypt';
+export * from './keccak256';