+++ packages/fuel-core/src/index.ts
@@ -1,5 +1,0 @@
-import path from 'path';
-
-const binPath = path.join(__dirname, '../fuel-core-binaries/fuel-core');
-
-export default binPath;