+++ packages/wallet/src/test-utils/seedTestWallet.ts
@@ -1,10 +1,11 @@
-import { randomBytes } from '@fuel-ts/keystore';
+import { randomBytes } from '@fuel-ts/crypto';
 import { coinQuantityfy, ScriptTransactionRequest } from '@fuel-ts/providers';
 import type { CoinQuantityLike } from '@fuel-ts/providers';
 
+import type { Account } from '../account';
 import { WalletUnlocked } from '../wallets';
 
-export const seedTestWallet = async (wallet: WalletUnlocked, quantities: CoinQuantityLike[]) => {
+export const seedTestWallet = async (wallet: Account, quantities: CoinQuantityLike[]) => {
   const genesisWallet = new WalletUnlocked(
     process.env.GENESIS_SECRET || randomBytes(32),
     wallet.provider