+++ apps/demo-typegen/CHANGELOG.md
@@ -11,3 +11,15 @@
 ## null
 
 ## null
+
+## null
+
+## null
+
+## null
+
+## null
+
+## null
+
+## null