+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @fuel-ts/docs
 
+## 0.42.1
+
+## 0.42.0
+
+### Minor Changes
+
+- use simulate instead of get on BaseInvocationScope class, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1179](https://github.com/FuelLabs/fuels-ts/pull/1179))
+
+### Patch Changes
+
+- add docs CI, match docs nav names to files, by [@sarahschwartz](https://github.com/sarahschwartz) (See [#1177](https://github.com/FuelLabs/fuels-ts/pull/1177))
+
+## 0.41.6
+
+## 0.41.5
+
+## 0.41.4
+
+### Patch Changes
+
+- Deprecate tx funding call option as all txs require a spendable input, by [@danielbate](https://github.com/danielbate) (See [#1136](https://github.com/FuelLabs/fuels-ts/pull/1136))
+
+## 0.41.3
+
 ## 0.41.2
 
 ## 0.41.1