+++ packages/providers/CHANGELOG.md
@@ -1,5 +1,37 @@
 # Change Log
 
+## 0.51.0
+
+### Minor Changes
+
+- upgrade fuel-core to 0.20.3, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1112](https://github.com/FuelLabs/fuels-ts/pull/1112))
+
+## 0.50.0
+
+### Minor Changes
+
+- improve transaction response, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1108](https://github.com/FuelLabs/fuels-ts/pull/1108))
+
+## 0.49.1
+
+## 0.49.0
+
+### Minor Changes
+
+- rename package keystore to crypto, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1140](https://github.com/FuelLabs/fuels-ts/pull/1140))
+
+### Patch Changes
+
+- 🐞 Fix predicate check on estimateTxDependencies, by [@luizstacio](https://github.com/luizstacio) (See [#1142](https://github.com/FuelLabs/fuels-ts/pull/1142))
+
+## 0.48.2
+
+### Patch Changes
+
+- 🐞 Fix predicate estimation for inputs with predicate field, by [@luizstacio](https://github.com/luizstacio) (See [#1130](https://github.com/FuelLabs/fuels-ts/pull/1130))
+
+## 0.48.1
+
 ## 0.48.0
 
 ### Minor Changes