+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 0.51.0
+
+## 0.50.0
+
+### Minor Changes
+
+- support encrypt and decrypt json wallets, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1041](https://github.com/FuelLabs/fuels-ts/pull/1041))
+
+### Patch Changes
+
+- Adding WASM integration for `@fuels/vm-asm`, by [@camsjams](https://github.com/camsjams) (See [#1164](https://github.com/FuelLabs/fuels-ts/pull/1164))
+
+## 0.49.1
+
+## 0.49.0
+
+### Minor Changes
+
+- rename package keystore to crypto, by [@Torres-ssf](https://github.com/Torres-ssf) (See [#1140](https://github.com/FuelLabs/fuels-ts/pull/1140))
+
+## 0.48.2
+
+## 0.48.1
+
 ## 0.48.0
 
 ### Patch Changes